Revert default build options changes
Revert 'USE_REDIS_DEFAULT' to be 'OFF'
Revert 'USE_IBVERBS_DEFAULT' to be 'OFF'
Revert 'BUILD_BENCHMARK' to be 'OFF'

Revert Yaniv's change. Yaniv changed the default values of the build
options instead of simply using environment variables propely.
Because of this change, Yaniv also had to change the target name
of the Gloo benchmark because is had conflicts with PyTorch when
building PyTorch from source.